Successful Blender cube and extensive comments on mesh loading
Added several useful models for testing, including a cube, a cuboid, and a fungus which only works some of the time.
Added extensive comments in ExternalForceDemo.cpp. These comments are about tetmodel generation and start to explain how tet models are extracted from files and used from there. More to come, as I'm currently trying to figure out why the armadillos are so squishy but the Blender cubes/fungus aren't