Made a squishy cube../bin/ExternalForceDemo ../bin/resources/scenes/GameLevelScene.json
GameLevelScene now contains a very squishy fungus and a slightly squishy cube. The cube kind of shakes and ripples, it's definitely a little squishy. Fantastic. Now we need the cube to be squishier. Currently it has 10 subdivisions, which was the maximum I could add in Blender with the simple subdivision modifier. But with a Blender script, we could do *way* more. We could even restrict the subdivisions to one face.