Made very squishy cube!
Made a very squishy cube in Blender by subdividing with 10 cuts, then selecting all the faces and subdividing 2 again. So 12 subdivisions total. Then exported and put it in GameLevelScene.json. It works pretty well, and the cube is really squishy - borderline liquid. It interacts beautifully with the character (rigid body) cube as well. It also runs at a rate of 1 second per second, which is probably a little too slow (we'll need to tone down the quality of the cube, maybe to 11 subdivisions, if we want lots of walls), but is funnily enough just the right speed for a real game.