SQUISHY WALLS WORK
Made a squishy wall model in Blender which really works. The wall is a cuboid that looks a bit like an airbag. Will need to fix it up later and make the outer faces - at least the one that is visible by the camera - more detailed, so that you don't get weird artifacts from the low-res mesh when it squishes