Convincing front and back wall
Added capability for invisible rigid bodies. Seems to work well but might not work in all cases - keep an eye out.
Add a convincing invisible front wall, with a visible back wall. Might need to play around with their various frictions/restitutions, etc. In addition, they might be a little close together right now. The distance between them allows for a satisfying tiny freedom of movement of the player in the z direction (i.e. the cube turns left or right a tiny amount, which makes the overall motion a little deeper and more satisfying). But if the game lags it might be that the walls are too close together, so we'll have to experiment.