Slightly better movement with A and D
Remapped movement so that left is A and right is D. Also has some structure in place for jumps with W in the future. Was a little tricky to set up the class properly so the keys are properly pressed and released, because there was no great system for releasing keys.
Also, remapped wireframe mode from w to i, and zoom from a to u. Was a little tricky to figure out where these things are, so looking at the changes in this commit may be useful in future.