Restructured Recoil
Structure of class for controlling forces wasn't great because force calculations would need to be repeated. Restructured so that ForceController is now called Thrusters, and only handles actually applying the recoil force. A higher class called PlayerController now handles the user input, calculating the recoil force and telling Thrusters to activate. This will mean that PlayerController can also tell Shooter to shoot a bullet, giving it the recoil force in the same way.
One issue: PlayerController is in ExternalForceDemo folder, and needs to know about MiniGL. However, this means that TimeStepController, which is in simulation, needs to know about a file in a specific demo, which means we can't run any other demos (which is a problem that has existed for a little while and probably doesn't matter). Better might be if PlayerController were in Simulation, and the MiniGL functions it needs were passed in. The code for that is in PlayerController mostly, but it's commented out.