Removed PlayerController MiniGL dependency
Move PlayerController to Simulation (from Demos folder) and removed its dependency on MiniGL to solve problem from previous commit. The necessary MiniGL functions are now passed into it.
The issue was that Simulation/TimeStepController was calling Demos/.../PlayerController,which meant that PBD needed to know about the demos... Now the player functionality is actulally part of the library.