Simple camera works
Simple camera follows player, using a log function to gradually move more slowly as it approaches a certain radius around the player.
Did a little bit of fine tuning until I found parameters I liked. Not sure that log is the right function for this though. I thought I'd want something that accelerates a little faster, but it actually does the trick pretty well.
In the future, we may have issues with visibility with the current camera - you can't always see in front of you super well. But this may be fixed with a bigger screen, and we can't know until the first level is designed.
In addition, when the game starts the camera jumps - probably because of some weird initialization of variables in the first update.
Finally, is the camera update() function called at the right time in the loop? Right now it's called at the very end of the step() loop, after the updating of the visualization meshes for whatever reason. I would've liked to put it in the MiniGL update loop, but that would force every demo to have a camera. Maybe that wouldn't be that bad actually - update the MiniGL module so that if you set up a camera, MiniGL calls update on it, and if you don't the camera is just static.